requirements.txt
chromadb==0.3.29
duckdb==0.8.1
langchain==0.0.288
langsmith==0.0.39
numpy==1.26.0
openai==0.27.6
pandas==2.0.3
pandasql==0.7.3
pypdf==3.16.1
python-dotenv
streamlit==1.27.0
flask
flask-cors
gunicorn
tiktoken
Flask-Mail
apscheduler
pytz
flask_jwt_extended
passlib
pymongo